pyproject.toml
[project]
name = "fftthing"
version = "0.1.0"
description = ""
authors = [
    {name = "dotkalm",email = "joel.holmberg@myriad.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "scipy (>=1.15.1,<2.0.0)",
    "img2wav (>=0.1.2,<0.2.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "openpyxl (>=3.1.5,<4.0.0)",
    "pillow (>=11.1.0,<12.0.0)",
    "svgwrite (>=1.4.3,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
